requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# AI and ML
anthropic==0.60.0
pandas==2.1.3
numpy==1.26.2

# Vector database
chromadb==0.4.18

# Additional utilities
aiofiles==23.2.1
typing-extensions==4.8.0